+++ e2e_example/test/common/utils.dart
@@ -180,11 +180,11 @@ Future<Null> get nextFailedBuild async {
   await _stdErrLines.firstWhere((line) => line.contains('Build: Failed after'));
 }
 
-Future<String> nextStdErrLine(String message) =>
-    _stdErrLines.firstWhere((line) => line.contains(message));
+Future<String> nextStdErrLine(String message) => _stdErrLines
+    .firstWhere((line) => line.contains(message));
 
-Future<String> nextStdOutLine(String message) =>
-    _stdOutLines.firstWhere((line) => line.contains(message));
+Future<String> nextStdOutLine(String message) => _stdOutLines
+    .firstWhere((line) => line.contains(message));
 
 /// Runs tests using the manual build script.
 Future<ProcessResult> runManualTests(